[wip] replace lsp-nvim-installer with mason.nvim
Replace lsp-nvim-installer with mason.nvim. Installed mason.nvim and
mason-lspconfig.nvim. The latter bridges some gaps between mason.nvim
and lspconfig. It's a work in progress. Installation for language
servers isn't working yet, likely because required apps like 'unzip' are
not installed in the Docker container. However, mason.nvim and
mason-lspconfig.nvim commands are working which is promising. There is
also a folder with language servers settings that might have to be
deleted in the future if I can't figure out how to configure language
servers individually with their individual .lua files.