[feat] add symbols-bar plugin
First attempt to find a replacement for Tagbar. I had used Tagbar in all
of my Vim IDE configurations but wanted to find a Lua implementation of
a tag overview for Neovim. As far as I know, symbols-bar is not based on
ctags so I will have to use it for a bit to find out if it's efficient
and if I like it. In its readme, symbols-bar is described as: "A tree
like view for symbols in Neovim using the Language Server Protocol.
Supports all your favourite languages."